+++ port_ocean/cli/commands/defaults/dock.py
@@ -2,11 +2,13 @@
 import click
 
 from inspect import getmembers
+
+from port_ocean.utils import load_module
 from .group import defaults
 from port_ocean.cli.commands.main import print_logo, console
 from port_ocean.ocean import Ocean
-from port_ocean.cli.defaults.initialize import initialize_defaults
-from port_ocean.run import _create_default_app, _load_module
+from port_ocean.core.defaults.initialize import initialize_defaults
+from port_ocean.bootstrap import create_default_app
 
 
 @defaults.command()
@@ -21,11 +23,11 @@ def dock(path: str) -> None:
 
     console.print("Unloading cargo at the dock... 📦🚢")
 
-    default_app = _create_default_app(path, False)
+    default_app = create_default_app(path)
 
     main_path = f"{path}/main.py" if path else "main.py"
 
-    app_module = _load_module(main_path)
+    app_module = load_module(main_path)
     app: Ocean = {name: item for name, item in getmembers(app_module)}.get(
         "app",
         default_app,
