+++ port_ocean/run.py
@@ -1,62 +1,25 @@
-import sys
-from importlib.util import spec_from_file_location, module_from_spec
-from inspect import getmembers, isclass
-from types import ModuleType
+from inspect import getmembers
 from typing import Type
+from pydantic import BaseModel
 
 import uvicorn
 
+from port_ocean.bootstrap import create_default_app
 from port_ocean.config.dynamic import default_config_factory
-from port_ocean.config.settings import LogLevelType, ApplicationSettings
-from port_ocean.core.integrations.base import BaseIntegration
+from port_ocean.config.settings import ApplicationSettings, LogLevelType
+from port_ocean.core.defaults.initialize import initialize_defaults
 from port_ocean.logger_setup import setup_logger
 from port_ocean.ocean import Ocean
-from port_ocean.cli.defaults.initialize import initialize_defaults
-from port_ocean.utils import get_spec_file
+from port_ocean.utils import get_spec_file, load_module
 
 
-def _get_base_integration_class_from_module(
-    module: ModuleType,
-) -> Type[BaseIntegration]:
-    for name, obj in getmembers(module):
-        if (
-            isclass(obj)
-            and type(obj) == type
-            and issubclass(obj, BaseIntegration)
-            and obj != BaseIntegration
-        ):
-            return obj
-
-    raise Exception(f"Failed to load integration from module: {module.__name__}")
-
-
-def _load_module(file_path: str) -> ModuleType:
-    spec = spec_from_file_location("module.name", file_path)
-    if spec is None or spec.loader is None:
-        raise Exception(f"Failed to load integration from path: {file_path}")
-
-    module = module_from_spec(spec)
-    spec.loader.exec_module(module)
-
-    return module
-
-
-def _create_default_app(
-    path: str | None = None, use_default_config_factory: bool = False
-) -> Ocean:
-    sys.path.append(".")
-    try:
-        integration_path = f"{path}/integration.py" if path else "integration.py"
-        module = _load_module(integration_path)
-        integration_class = _get_base_integration_class_from_module(module)
-    except Exception:
-        integration_class = None
-
+def _get_default_config_factory() -> None | Type[BaseModel]:
     spec = get_spec_file()
     config_factory = None
-    if spec is not None and use_default_config_factory:
+    if spec is not None:
         config_factory = default_config_factory(spec.get("configurations", []))
-    return Ocean(integration_class=integration_class, config_factory=config_factory)
+
+    return config_factory
 
 
 def run(
@@ -68,10 +31,11 @@ def run(
     application_settings = ApplicationSettings(log_level=log_level, port=port)
 
     setup_logger(application_settings.log_level)
-    default_app = _create_default_app(path)
+    config_factory = _get_default_config_factory()
+    default_app = create_default_app(path, config_factory)
 
     main_path = f"{path}/main.py" if path else "main.py"
-    app_module = _load_module(main_path)
+    app_module = load_module(main_path)
     app: Ocean = {name: item for name, item in getmembers(app_module)}.get(
         "app", default_app
     )
